+++ app.go
@@ -2,27 +2,22 @@ package main
 
 import (
 	crand "crypto/rand"
-	"crypto/sha1"
-	"database/sql"
 	"encoding/binary"
 	"fmt"
 	"html/template"
 	"io"
-	"io/ioutil"
 	"log"
 	"math/rand"
 	"net/http"
 	"os"
-	"strconv"
-	"strings"
 	"time"
 
-	"github.com/go-sql-driver/mysql"
 	"github.com/gorilla/sessions"
 	"github.com/jmoiron/sqlx"
 	"github.com/labstack/echo"
 	"github.com/labstack/echo-contrib/session"
 	"github.com/labstack/echo/middleware"
+
 	_ "net/http/pprof"
 )
 
@@ -89,142 +84,6 @@ func init() {
 	db.SetConnMaxLifetime(5 * time.Minute)
 	log.Printf("Succeeded to connect db.")
 }
-
-type User struct {
-	ID          int64     `json:"-" db:"id"`
-	Name        string    `json:"name" db:"name"`
-	Salt        string    `json:"-" db:"salt"`
-	Password    string    `json:"-" db:"password"`
-	DisplayName string    `json:"display_name" db:"display_name"`
-	AvatarIcon  string    `json:"avatar_icon" db:"avatar_icon"`
-	CreatedAt   time.Time `json:"-" db:"created_at"`
-}
-
-func getUser(userID int64) (*User, error) {
-	u := User{}
-	if err := db.Get(&u, "SELECT * FROM user WHERE id = ?", userID); err != nil {
-		if err == sql.ErrNoRows {
-			return nil, nil
-		}
-		return nil, err
-	}
-	return &u, nil
-}
-
-func addMessage(channelID, userID int64, content string) (int64, error) {
-	res, err := db.Exec(
-		"INSERT INTO message (channel_id, user_id, content, created_at) VALUES (?, ?, ?, NOW())",
-		channelID, userID, content)
-	if err != nil {
-		return 0, err
-	}
-	return res.LastInsertId()
-}
-
-type Message struct {
-	ID        int64     `db:"id"`
-	ChannelID int64     `db:"channel_id"`
-	UserID    int64     `db:"user_id"`
-	Content   string    `db:"content"`
-	CreatedAt time.Time `db:"created_at"`
-}
-
-type MessageAndUser struct {
-	ID        int64     `db:"id"`
-	ChannelID int64     `db:"channel_id"`
-	UserID    int64     `db:"user_id"`
-	Content   string    `db:"content"`
-	CreatedAt time.Time `db:"created_at"`
-
-	UserPK          int64     `json:"-" db:"user_pk"`
-	UserName        string    `json:"name" db:"name"`
-	UserSalt        string    `json:"-" db:"salt"`
-	UserPassword    string    `json:"-" db:"password"`
-	UserDisplayName string    `json:"display_name" db:"display_name"`
-	UserAvatarIcon  string    `json:"avatar_icon" db:"avatar_icon"`
-	UserCreatedAt   time.Time `json:"-" db:"user_created_at"`
-}
-
-func queryMessages(chanID, lastID int64) ([]MessageAndUser, error) {
-	msgs := []MessageAndUser{}
-	err := db.Select(&msgs, "SELECT m.*, u.id as user_pk, u.name, u.salt, u.password, u.display_name, u.avatar_icon, u.created_at as user_created_at FROM message AS m JOIN user AS u ON m.user_id = u.id WHERE m.id > ? AND m.channel_id = ? ORDER BY m.id DESC LIMIT 100",
-		lastID, chanID)
-	return msgs, err
-}
-
-func sessUserID(c echo.Context) int64 {
-	sess, _ := session.Get("session", c)
-	var userID int64
-	if x, ok := sess.Values["user_id"]; ok {
-		userID, _ = x.(int64)
-	}
-	return userID
-}
-
-func sessSetUserID(c echo.Context, id int64) {
-	sess, _ := session.Get("session", c)
-	sess.Options = &sessions.Options{
-		HttpOnly: true,
-		MaxAge:   360000,
-	}
-	sess.Values["user_id"] = id
-	sess.Save(c.Request(), c.Response())
-}
-
-func ensureLogin(c echo.Context) (*User, error) {
-	var user *User
-	var err error
-
-	userID := sessUserID(c)
-	if userID == 0 {
-		goto redirect
-	}
-
-	user, err = getUser(userID)
-	if err != nil {
-		return nil, err
-	}
-	if user == nil {
-		sess, _ := session.Get("session", c)
-		delete(sess.Values, "user_id")
-		sess.Save(c.Request(), c.Response())
-		goto redirect
-	}
-	return user, nil
-
-redirect:
-	c.Redirect(http.StatusSeeOther, "/login")
-	return nil, nil
-}
-
-const LettersAndDigits = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
-
-func randomString(n int) string {
-	b := make([]byte, n)
-	z := len(LettersAndDigits)
-
-	for i := 0; i < n; i++ {
-		b[i] = LettersAndDigits[rand.Intn(z)]
-	}
-	return string(b)
-}
-
-func register(name, password string) (int64, error) {
-	salt := randomString(20)
-	digest := fmt.Sprintf("%x", sha1.Sum([]byte(salt+password)))
-
-	res, err := db.Exec(
-		"INSERT INTO user (name, salt, password, display_name, avatar_icon, created_at)"+
-			" VALUES (?, ?, ?, ?, ?, NOW())",
-		name, salt, digest, name, "default.png")
-	if err != nil {
-		return 0, err
-	}
-	return res.LastInsertId()
-}
-
-// request handlers
-
 func getInitialize(c echo.Context) error {
 	db.MustExec("DELETE FROM user WHERE id > 1000")
 	db.MustExec("DELETE FROM image WHERE id > 1001")
@@ -243,463 +102,6 @@ func getIndex(c echo.Context) error {
 	return c.Render(http.StatusOK, "index", map[string]interface{}{
 		"ChannelID": nil,
 	})
-}
-
-type ChannelInfo struct {
-	ID          int64     `db:"id"`
-	Name        string    `db:"name"`
-	Description string    `db:"description"`
-	UpdatedAt   time.Time `db:"updated_at"`
-	CreatedAt   time.Time `db:"created_at"`
-}
-
-func getChannel(c echo.Context) error {
-	user, err := ensureLogin(c)
-	if user == nil {
-		return err
-	}
-	cID, err := strconv.Atoi(c.Param("channel_id"))
-	if err != nil {
-		return err
-	}
-	channels := []ChannelInfo{}
-	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
-	if err != nil {
-		return err
-	}
-
-	var desc string
-	for _, ch := range channels {
-		if ch.ID == int64(cID) {
-			desc = ch.Description
-			break
-		}
-	}
-	return c.Render(http.StatusOK, "channel", map[string]interface{}{
-		"ChannelID":   cID,
-		"Channels":    channels,
-		"User":        user,
-		"Description": desc,
-	})
-}
-
-func getRegister(c echo.Context) error {
-	return c.Render(http.StatusOK, "register", map[string]interface{}{
-		"ChannelID": 0,
-		"Channels":  []ChannelInfo{},
-		"User":      nil,
-	})
-}
-
-func postRegister(c echo.Context) error {
-	name := c.FormValue("name")
-	pw := c.FormValue("password")
-	if name == "" || pw == "" {
-		return ErrBadReqeust
-	}
-	userID, err := register(name, pw)
-	if err != nil {
-		if merr, ok := err.(*mysql.MySQLError); ok {
-			if merr.Number == 1062 { // Duplicate entry xxxx for key zzzz
-				return c.NoContent(http.StatusConflict)
-			}
-		}
-		return err
-	}
-	sessSetUserID(c, userID)
-	return c.Redirect(http.StatusSeeOther, "/")
-}
-
-func getLogin(c echo.Context) error {
-	return c.Render(http.StatusOK, "login", map[string]interface{}{
-		"ChannelID": 0,
-		"Channels":  []ChannelInfo{},
-		"User":      nil,
-	})
-}
-
-func postLogin(c echo.Context) error {
-	name := c.FormValue("name")
-	pw := c.FormValue("password")
-	if name == "" || pw == "" {
-		return ErrBadReqeust
-	}
-
-	var user User
-	err := db.Get(&user, "SELECT * FROM user WHERE name = ?", name)
-	if err == sql.ErrNoRows {
-		return echo.ErrForbidden
-	} else if err != nil {
-		return err
-	}
-
-	digest := fmt.Sprintf("%x", sha1.Sum([]byte(user.Salt+pw)))
-	if digest != user.Password {
-		return echo.ErrForbidden
-	}
-	sessSetUserID(c, user.ID)
-	return c.Redirect(http.StatusSeeOther, "/")
-}
-
-func getLogout(c echo.Context) error {
-	sess, _ := session.Get("session", c)
-	delete(sess.Values, "user_id")
-	sess.Save(c.Request(), c.Response())
-	return c.Redirect(http.StatusSeeOther, "/")
-}
-
-func postMessage(c echo.Context) error {
-	user, err := ensureLogin(c)
-	if user == nil {
-		return err
-	}
-
-	message := c.FormValue("message")
-	if message == "" {
-		return echo.ErrForbidden
-	}
-
-	var chanID int64
-	if x, err := strconv.Atoi(c.FormValue("channel_id")); err != nil {
-		return echo.ErrForbidden
-	} else {
-		chanID = int64(x)
-	}
-
-	if _, err := addMessage(chanID, user.ID, message); err != nil {
-		return err
-	}
-
-	return c.NoContent(204)
-}
-
-func jsonifyMessage(m MessageAndUser) (map[string]interface{}, error) {
-	u := User{
-		m.UserPK,
-		m.UserName,
-		m.UserSalt,
-		m.UserPassword,
-		m.UserDisplayName,
-		m.UserAvatarIcon,
-		m.UserCreatedAt,
-	}
-
-	r := make(map[string]interface{})
-	r["id"] = m.ID
-	r["user"] = u
-	r["date"] = m.CreatedAt.Format("2006/01/02 15:04:05")
-	r["content"] = m.Content
-	return r, nil
-}
-
-func getMessage(c echo.Context) error {
-	userID := sessUserID(c)
-	if userID == 0 {
-		return c.NoContent(http.StatusForbidden)
-	}
-
-	chanID, err := strconv.ParseInt(c.QueryParam("channel_id"), 10, 64)
-	if err != nil {
-		return err
-	}
-	lastID, err := strconv.ParseInt(c.QueryParam("last_message_id"), 10, 64)
-	if err != nil {
-		return err
-	}
-
-	messages, err := queryMessages(chanID, lastID)
-	if err != nil {
-		return err
-	}
-
-	response := make([]map[string]interface{}, 0)
-	for i := len(messages) - 1; i >= 0; i-- {
-		m := messages[i]
-		r, err := jsonifyMessage(m)
-		if err != nil {
-			return err
-		}
-		response = append(response, r)
-	}
-
-	if len(messages) > 0 {
-		r := Read{
-			UserID:    userID,
-			ChannelID: chanID,
-		}
-		lastRead[r] = messages[0].ID
-	}
-
-	return c.JSON(http.StatusOK, response)
-}
-
-func queryChannels() ([]int64, error) {
-	res := []int64{}
-	err := db.Select(&res, "SELECT id FROM channel")
-	return res, err
-}
-
-func fetchUnread(c echo.Context) error {
-	userID := sessUserID(c)
-	if userID == 0 {
-		return c.NoContent(http.StatusForbidden)
-	}
-
-	time.Sleep(time.Second)
-
-	channels, err := queryChannels()
-	if err != nil {
-		return err
-	}
-
-	resp := []map[string]interface{}{}
-
-	for _, chID := range channels {
-		lastID, flag := lastRead[Read{
-			UserID:    userID,
-			ChannelID: chID,
-		}]
-
-		var cnt int64
-		if flag {
-			err = db.Get(&cnt,
-				"SELECT COUNT(*) as cnt FROM message WHERE channel_id = ? AND ? < id",
-				chID, lastID)
-		} else {
-			err = db.Get(&cnt,
-				"SELECT COUNT(*) as cnt FROM message WHERE channel_id = ?",
-				chID)
-		}
-		if err != nil {
-			return err
-		}
-		r := map[string]interface{}{
-			"channel_id": chID,
-			"unread":     cnt}
-		resp = append(resp, r)
-	}
-
-	return c.JSON(http.StatusOK, resp)
-}
-
-func getHistory(c echo.Context) error {
-	chID, err := strconv.ParseInt(c.Param("channel_id"), 10, 64)
-	if err != nil || chID <= 0 {
-		return ErrBadReqeust
-	}
-
-	user, err := ensureLogin(c)
-	if user == nil {
-		return err
-	}
-
-	var page int64
-	pageStr := c.QueryParam("page")
-	if pageStr == "" {
-		page = 1
-	} else {
-		page, err = strconv.ParseInt(pageStr, 10, 64)
-		if err != nil || page < 1 {
-			return ErrBadReqeust
-		}
-	}
-
-	const N = 20
-	var cnt int64
-	err = db.Get(&cnt, "SELECT COUNT(*) as cnt FROM message WHERE channel_id = ?", chID)
-	if err != nil {
-		return err
-	}
-	maxPage := int64(cnt+N-1) / N
-	if maxPage == 0 {
-		maxPage = 1
-	}
-	if page > maxPage {
-		return ErrBadReqeust
-	}
-
-	messages := []MessageAndUser{}
-	err = db.Select(&messages,
-		"SELECT m.*,"+
-			" u.id AS user_pk, u.name, u.salt, u.password, u.display_name, u.avatar_icon, u.created_at AS user_created_at"+
-			" FROM message AS m"+
-			" LEFT JOIN user AS u"+
-			" ON m.user_id = u.id"+
-			" WHERE m.channel_id = ?"+
-			" ORDER BY m.id DESC LIMIT ? OFFSET ?",
-		chID, N, (page-1)*N)
-	if err != nil {
-		return err
-	}
-
-	mjson := make([]map[string]interface{}, 0)
-	for i := len(messages) - 1; i >= 0; i-- {
-		r, err := jsonifyMessage(messages[i])
-		if err != nil {
-			return err
-		}
-		mjson = append(mjson, r)
-	}
-
-	channels := []ChannelInfo{}
-	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
-	if err != nil {
-		return err
-	}
-
-	return c.Render(http.StatusOK, "history", map[string]interface{}{
-		"ChannelID": chID,
-		"Channels":  channels,
-		"Messages":  mjson,
-		"MaxPage":   maxPage,
-		"Page":      page,
-		"User":      user,
-	})
-}
-
-func getProfile(c echo.Context) error {
-	self, err := ensureLogin(c)
-	if self == nil {
-		return err
-	}
-
-	channels := []ChannelInfo{}
-	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
-	if err != nil {
-		return err
-	}
-
-	userName := c.Param("user_name")
-	var other User
-	err = db.Get(&other, "SELECT * FROM user WHERE name = ?", userName)
-	if err == sql.ErrNoRows {
-		return echo.ErrNotFound
-	}
-	if err != nil {
-		return err
-	}
-
-	return c.Render(http.StatusOK, "profile", map[string]interface{}{
-		"ChannelID":   0,
-		"Channels":    channels,
-		"User":        self,
-		"Other":       other,
-		"SelfProfile": self.ID == other.ID,
-	})
-}
-
-func getAddChannel(c echo.Context) error {
-	self, err := ensureLogin(c)
-	if self == nil {
-		return err
-	}
-
-	channels := []ChannelInfo{}
-	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
-	if err != nil {
-		return err
-	}
-
-	return c.Render(http.StatusOK, "add_channel", map[string]interface{}{
-		"ChannelID": 0,
-		"Channels":  channels,
-		"User":      self,
-	})
-}
-
-func postAddChannel(c echo.Context) error {
-	self, err := ensureLogin(c)
-	if self == nil {
-		return err
-	}
-
-	name := c.FormValue("name")
-	desc := c.FormValue("description")
-	if name == "" || desc == "" {
-		return ErrBadReqeust
-	}
-
-	res, err := db.Exec(
-		"INSERT INTO channel (name, description, updated_at, created_at) VALUES (?, ?, NOW(), NOW())",
-		name, desc)
-	if err != nil {
-		return err
-	}
-	lastID, _ := res.LastInsertId()
-	return c.Redirect(http.StatusSeeOther,
-		fmt.Sprintf("/channel/%v", lastID))
-}
-
-func postProfile(c echo.Context) error {
-	self, err := ensureLogin(c)
-	if self == nil {
-		return err
-	}
-
-	avatarName := ""
-	var avatarData []byte
-
-	if fh, err := c.FormFile("avatar_icon"); err == http.ErrMissingFile {
-		// no file upload
-	} else if err != nil {
-		return err
-	} else {
-		dotPos := strings.LastIndexByte(fh.Filename, '.')
-		if dotPos < 0 {
-			return ErrBadReqeust
-		}
-		ext := fh.Filename[dotPos:]
-		switch ext {
-		case ".jpg", ".jpeg", ".png", ".gif":
-			break
-		default:
-			return ErrBadReqeust
-		}
-
-		file, err := fh.Open()
-		if err != nil {
-			return err
-		}
-		avatarData, _ = ioutil.ReadAll(file)
-		file.Close()
-
-		if len(avatarData) > avatarMaxBytes {
-			return ErrBadReqeust
-		}
-
-		avatarName = fmt.Sprintf("%x%s", sha1.Sum(avatarData), ext)
-	}
-
-	if avatarName != "" && len(avatarData) > 0 {
-
-		ioutil.WriteFile("/var/www/icons/"+avatarName, avatarData, 0644)
-
-		_, err = db.Exec("UPDATE user SET avatar_icon = ? WHERE id = ?", avatarName, self.ID)
-		if err != nil {
-			return err
-		}
-	}
-
-	if name := c.FormValue("display_name"); name != "" {
-		_, err := db.Exec("UPDATE user SET display_name = ? WHERE id = ?", name, self.ID)
-		if err != nil {
-			return err
-		}
-	}
-
-	return c.Redirect(http.StatusSeeOther, "/")
-}
-
-func tAdd(a, b int64) int64 {
-	return a + b
-}
-
-func tRange(a, b int64) []int64 {
-	r := make([]int64, b-a+1)
-	for i := int64(0); i <= (b - a); i++ {
-		r[i] = a + i
-	}
-	return r
 }
 
 func main() {
