+++ auth.go
@@ -1,0 +1,149 @@
+package main
+
+import (
+	"crypto/sha1"
+	"database/sql"
+	"fmt"
+	"github.com/go-sql-driver/mysql"
+	"github.com/gorilla/sessions"
+	"github.com/labstack/echo"
+	"github.com/labstack/echo-contrib/session"
+	"math/rand"
+	"net/http"
+)
+
+func sessUserID(c echo.Context) int64 {
+	sess, _ := session.Get("session", c)
+	var userID int64
+	if x, ok := sess.Values["user_id"]; ok {
+		userID, _ = x.(int64)
+	}
+	return userID
+}
+
+func sessSetUserID(c echo.Context, id int64) {
+	sess, _ := session.Get("session", c)
+	sess.Options = &sessions.Options{
+		HttpOnly: true,
+		MaxAge:   360000,
+	}
+	sess.Values["user_id"] = id
+	sess.Save(c.Request(), c.Response())
+}
+
+func ensureLogin(c echo.Context) (*User, error) {
+	var user *User
+	var err error
+
+	userID := sessUserID(c)
+	if userID == 0 {
+		goto redirect
+	}
+
+	user, err = getUser(userID)
+	if err != nil {
+		return nil, err
+	}
+	if user == nil {
+		sess, _ := session.Get("session", c)
+		delete(sess.Values, "user_id")
+		sess.Save(c.Request(), c.Response())
+		goto redirect
+	}
+	return user, nil
+
+redirect:
+	c.Redirect(http.StatusSeeOther, "/login")
+	return nil, nil
+}
+
+const LettersAndDigits = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
+
+func randomString(n int) string {
+	b := make([]byte, n)
+	z := len(LettersAndDigits)
+
+	for i := 0; i < n; i++ {
+		b[i] = LettersAndDigits[rand.Intn(z)]
+	}
+	return string(b)
+}
+
+func register(name, password string) (int64, error) {
+	salt := randomString(20)
+	digest := fmt.Sprintf("%x", sha1.Sum([]byte(salt+password)))
+
+	res, err := db.Exec(
+		"INSERT INTO user (name, salt, password, display_name, avatar_icon, created_at)"+
+			" VALUES (?, ?, ?, ?, ?, NOW())",
+		name, salt, digest, name, "default.png")
+	if err != nil {
+		return 0, err
+	}
+	return res.LastInsertId()
+}
+
+func getRegister(c echo.Context) error {
+	return c.Render(http.StatusOK, "register", map[string]interface{}{
+		"ChannelID": 0,
+		"Channels":  []ChannelInfo{},
+		"User":      nil,
+	})
+}
+
+func postRegister(c echo.Context) error {
+	name := c.FormValue("name")
+	pw := c.FormValue("password")
+	if name == "" || pw == "" {
+		return ErrBadReqeust
+	}
+	userID, err := register(name, pw)
+	if err != nil {
+		if merr, ok := err.(*mysql.MySQLError); ok {
+			if merr.Number == 1062 { // Duplicate entry xxxx for key zzzz
+				return c.NoContent(http.StatusConflict)
+			}
+		}
+		return err
+	}
+	sessSetUserID(c, userID)
+	return c.Redirect(http.StatusSeeOther, "/")
+}
+
+func getLogin(c echo.Context) error {
+	return c.Render(http.StatusOK, "login", map[string]interface{}{
+		"ChannelID": 0,
+		"Channels":  []ChannelInfo{},
+		"User":      nil,
+	})
+}
+
+func postLogin(c echo.Context) error {
+	name := c.FormValue("name")
+	pw := c.FormValue("password")
+	if name == "" || pw == "" {
+		return ErrBadReqeust
+	}
+
+	var user User
+	err := db.Get(&user, "SELECT * FROM user WHERE name = ?", name)
+	if err == sql.ErrNoRows {
+		return echo.ErrForbidden
+	} else if err != nil {
+		return err
+	}
+
+	digest := fmt.Sprintf("%x", sha1.Sum([]byte(user.Salt+pw)))
+	if digest != user.Password {
+		return echo.ErrForbidden
+	}
+	sessSetUserID(c, user.ID)
+	return c.Redirect(http.StatusSeeOther, "/")
+}
+
+func getLogout(c echo.Context) error {
+	sess, _ := session.Get("session", c)
+	delete(sess.Values, "user_id")
+	sess.Save(c.Request(), c.Response())
+	return c.Redirect(http.StatusSeeOther, "/")
+}
