+++ handler.go
@@ -1,0 +1,372 @@
+package main
+
+import (
+	"crypto/sha1"
+	"database/sql"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/labstack/echo"
+)
+
+func getChannel(c echo.Context) error {
+	user, err := ensureLogin(c)
+	if user == nil {
+		return err
+	}
+	cID, err := strconv.Atoi(c.Param("channel_id"))
+	if err != nil {
+		return err
+	}
+	channels := []ChannelInfo{}
+	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
+	if err != nil {
+		return err
+	}
+
+	var desc string
+	for _, ch := range channels {
+		if ch.ID == int64(cID) {
+			desc = ch.Description
+			break
+		}
+	}
+	return c.Render(http.StatusOK, "channel", map[string]interface{}{
+		"ChannelID":   cID,
+		"Channels":    channels,
+		"User":        user,
+		"Description": desc,
+	})
+}
+
+func getMessage(c echo.Context) error {
+	userID := sessUserID(c)
+	if userID == 0 {
+		return c.NoContent(http.StatusForbidden)
+	}
+
+	chanID, err := strconv.ParseInt(c.QueryParam("channel_id"), 10, 64)
+	if err != nil {
+		return err
+	}
+	lastID, err := strconv.ParseInt(c.QueryParam("last_message_id"), 10, 64)
+	if err != nil {
+		return err
+	}
+
+	messages, err := queryMessages(chanID, lastID)
+	if err != nil {
+		return err
+	}
+
+	response := make([]map[string]interface{}, 0)
+	for i := len(messages) - 1; i >= 0; i-- {
+		m := messages[i]
+		r, err := jsonifyMessage(m)
+		if err != nil {
+			return err
+		}
+		response = append(response, r)
+	}
+
+	if len(messages) > 0 {
+		r := Read{
+			UserID:    userID,
+			ChannelID: chanID,
+		}
+		lastRead[r] = messages[0].ID
+	}
+
+	return c.JSON(http.StatusOK, response)
+}
+
+func postMessage(c echo.Context) error {
+	user, err := ensureLogin(c)
+	if user == nil {
+		return err
+	}
+
+	message := c.FormValue("message")
+	if message == "" {
+		return echo.ErrForbidden
+	}
+
+	var chanID int64
+	if x, err := strconv.Atoi(c.FormValue("channel_id")); err != nil {
+		return echo.ErrForbidden
+	} else {
+		chanID = int64(x)
+	}
+
+	if _, err := addMessage(chanID, user.ID, message); err != nil {
+		return err
+	}
+
+	return c.NoContent(204)
+}
+
+func fetchUnread(c echo.Context) error {
+	userID := sessUserID(c)
+	if userID == 0 {
+		return c.NoContent(http.StatusForbidden)
+	}
+
+	time.Sleep(time.Second)
+
+	channels, err := queryChannels()
+	if err != nil {
+		return err
+	}
+
+	resp := []map[string]interface{}{}
+
+	for _, chID := range channels {
+		lastID, flag := lastRead[Read{
+			UserID:    userID,
+			ChannelID: chID,
+		}]
+
+		var cnt int64
+		if flag {
+			err = db.Get(&cnt,
+				"SELECT COUNT(*) as cnt FROM message WHERE channel_id = ? AND ? < id",
+				chID, lastID)
+		} else {
+			err = db.Get(&cnt,
+				"SELECT COUNT(*) as cnt FROM message WHERE channel_id = ?",
+				chID)
+		}
+		if err != nil {
+			return err
+		}
+		r := map[string]interface{}{
+			"channel_id": chID,
+			"unread":     cnt}
+		resp = append(resp, r)
+	}
+
+	return c.JSON(http.StatusOK, resp)
+}
+
+func getHistory(c echo.Context) error {
+	chID, err := strconv.ParseInt(c.Param("channel_id"), 10, 64)
+	if err != nil || chID <= 0 {
+		return ErrBadReqeust
+	}
+
+	user, err := ensureLogin(c)
+	if user == nil {
+		return err
+	}
+
+	var page int64
+	pageStr := c.QueryParam("page")
+	if pageStr == "" {
+		page = 1
+	} else {
+		page, err = strconv.ParseInt(pageStr, 10, 64)
+		if err != nil || page < 1 {
+			return ErrBadReqeust
+		}
+	}
+
+	const N = 20
+	var cnt int64
+	err = db.Get(&cnt, "SELECT COUNT(*) as cnt FROM message WHERE channel_id = ?", chID)
+	if err != nil {
+		return err
+	}
+	maxPage := int64(cnt+N-1) / N
+	if maxPage == 0 {
+		maxPage = 1
+	}
+	if page > maxPage {
+		return ErrBadReqeust
+	}
+
+	messages := []MessageAndUser{}
+	err = db.Select(&messages,
+		"SELECT m.*,"+
+			" u.id AS user_pk, u.name, u.salt, u.password, u.display_name, u.avatar_icon, u.created_at AS user_created_at"+
+			" FROM message AS m"+
+			" LEFT JOIN user AS u"+
+			" ON m.user_id = u.id"+
+			" WHERE m.channel_id = ?"+
+			" ORDER BY m.id DESC LIMIT ? OFFSET ?",
+		chID, N, (page-1)*N)
+	if err != nil {
+		return err
+	}
+
+	mjson := make([]map[string]interface{}, 0)
+	for i := len(messages) - 1; i >= 0; i-- {
+		r, err := jsonifyMessage(messages[i])
+		if err != nil {
+			return err
+		}
+		mjson = append(mjson, r)
+	}
+
+	channels := []ChannelInfo{}
+	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
+	if err != nil {
+		return err
+	}
+
+	return c.Render(http.StatusOK, "history", map[string]interface{}{
+		"ChannelID": chID,
+		"Channels":  channels,
+		"Messages":  mjson,
+		"MaxPage":   maxPage,
+		"Page":      page,
+		"User":      user,
+	})
+}
+
+func getProfile(c echo.Context) error {
+	self, err := ensureLogin(c)
+	if self == nil {
+		return err
+	}
+
+	channels := []ChannelInfo{}
+	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
+	if err != nil {
+		return err
+	}
+
+	userName := c.Param("user_name")
+	var other User
+	err = db.Get(&other, "SELECT * FROM user WHERE name = ?", userName)
+	if err == sql.ErrNoRows {
+		return echo.ErrNotFound
+	}
+	if err != nil {
+		return err
+	}
+
+	return c.Render(http.StatusOK, "profile", map[string]interface{}{
+		"ChannelID":   0,
+		"Channels":    channels,
+		"User":        self,
+		"Other":       other,
+		"SelfProfile": self.ID == other.ID,
+	})
+}
+
+func getAddChannel(c echo.Context) error {
+	self, err := ensureLogin(c)
+	if self == nil {
+		return err
+	}
+
+	channels := []ChannelInfo{}
+	err = db.Select(&channels, "SELECT * FROM channel ORDER BY id")
+	if err != nil {
+		return err
+	}
+
+	return c.Render(http.StatusOK, "add_channel", map[string]interface{}{
+		"ChannelID": 0,
+		"Channels":  channels,
+		"User":      self,
+	})
+}
+
+func postAddChannel(c echo.Context) error {
+	self, err := ensureLogin(c)
+	if self == nil {
+		return err
+	}
+
+	name := c.FormValue("name")
+	desc := c.FormValue("description")
+	if name == "" || desc == "" {
+		return ErrBadReqeust
+	}
+
+	res, err := db.Exec(
+		"INSERT INTO channel (name, description, updated_at, created_at) VALUES (?, ?, NOW(), NOW())",
+		name, desc)
+	if err != nil {
+		return err
+	}
+	lastID, _ := res.LastInsertId()
+	return c.Redirect(http.StatusSeeOther,
+		fmt.Sprintf("/channel/%v", lastID))
+}
+
+func postProfile(c echo.Context) error {
+	self, err := ensureLogin(c)
+	if self == nil {
+		return err
+	}
+
+	avatarName := ""
+	var avatarData []byte
+
+	if fh, err := c.FormFile("avatar_icon"); err == http.ErrMissingFile {
+		// no file upload
+	} else if err != nil {
+		return err
+	} else {
+		dotPos := strings.LastIndexByte(fh.Filename, '.')
+		if dotPos < 0 {
+			return ErrBadReqeust
+		}
+		ext := fh.Filename[dotPos:]
+		switch ext {
+		case ".jpg", ".jpeg", ".png", ".gif":
+			break
+		default:
+			return ErrBadReqeust
+		}
+
+		file, err := fh.Open()
+		if err != nil {
+			return err
+		}
+		avatarData, _ = ioutil.ReadAll(file)
+		file.Close()
+
+		if len(avatarData) > avatarMaxBytes {
+			return ErrBadReqeust
+		}
+
+		avatarName = fmt.Sprintf("%x%s", sha1.Sum(avatarData), ext)
+	}
+
+	if avatarName != "" && len(avatarData) > 0 {
+
+		ioutil.WriteFile("/var/www/icons/"+avatarName, avatarData, 0644)
+
+		_, err = db.Exec("UPDATE user SET avatar_icon = ? WHERE id = ?", avatarName, self.ID)
+		if err != nil {
+			return err
+		}
+	}
+
+	if name := c.FormValue("display_name"); name != "" {
+		_, err := db.Exec("UPDATE user SET display_name = ? WHERE id = ?", name, self.ID)
+		if err != nil {
+			return err
+		}
+	}
+
+	return c.Redirect(http.StatusSeeOther, "/")
+}
+
+func getUser(userID int64) (*User, error) {
+	u := User{}
+	if err := db.Get(&u, "SELECT * FROM user WHERE id = ?", userID); err != nil {
+		if err == sql.ErrNoRows {
+			return nil, nil
+		}
+		return nil, err
+	}
+	return &u, nil
+}
